--- app/.env
@@ -1,2 +1,0 @@
-# VITE_SERVER_BASE_URL = "http://localhost:8002"
-VITE_SERVER_BASE_URL = "https://steady-invest-involve-pn.trycloudflare.com"
