+++ app/src/types/interview.ts
@@ -1,0 +1,10 @@
+export interface InterviewState {
+  isConnected: boolean;
+  isCallActive: boolean;
+  connectionState: RTCPeerConnectionState;
+}
+
+export interface WebRTCResponse {
+  sdp: string;
+  type: string;
+}
